Added ItemDetailsSearchesView and fixed ItemDetailView
